[Driver][SYCL] Match up -device_options with -device for AOT GPU (#15678)
The driver will add additional -ftarget-register-alloc-mode values to
the spir64_gen AOT call to ocloc. These additional options are applied
when a user specifies any of the accepted PVC values that can be handled
by ocloc. These values can be 'pvc' or associated hex or version values.

Match up the -device_options arg value with the -device arg that is
passed on the command line. ocloc expects these to match in order to
properly associate the option given after -device_options.