[NFC][SYCL] Inline/drop `is_genint_v` type trait (#15707)
It's only used in a single header for extra filtering of `cl_int` and
`cl_float`. We don't need the entirety of "generic" integer types to do
that.

Note `(Dims > 0) && ` that I'm dropping here seemed to have no effect as
`IsValidCoordDataT` helper ensures `Dims` can't be non-positive. Also,
it would be a compiler error and not an SFINAE-based selection because
of that helper anyway.