[Driver][SYCL] Fix crash with empty -fsycl-targets option (#15766)
When -fsycl-targets= -fsycl-targets=arg is passed on the command line,
the empty -fsycl-targets= is processed and crashes due to the empty
argument. Update the specific option parsing to use 'getLastArg' to
avoid this situation. When the last arg has an empty argument, the
driver will diagnose this situation and emit a friendly diagnostic.