[SYCL][NFC] Introduce unit-tests for `ONEAPI_DEVICE_SELECTOR`
There are E2E tests which are essentially testing the correctness of
`ONEAPI_DEVICE_SELECTOR` env variable testing, but that can be done in
way more efficient way which is unit-tests.

Note that this PR does not aim to cover `ONEAPI_DEVICE_SELECTOR` parsing
with unit-tests in full, but rather aims to replace the most obvious
candidates from E2E tests with unit-tests, providing some foundation for
expanding unit-tests in follow-up PRs.