[Driver][SYCL] Address issue with code splitting and FPGA Archives
When enabling device code splitting with FPGA archive generation
(-fsycl-link) the resulting device images were not packed correctly.  We
were archiving the resulting -batch wrapped image, which in turn would
not be extracted properly when being consumed.

To address this, update how the device images are created for insertion
into the archive to be individual split binaries as opposed to using the
wrapper -batch behavior.  This is performed for both the early and image
type archives.