[SYCL][Docs][Joint matrix] Add overloads and restrictions for the offset load store (#15499)
- Add missing restriction on the stride of the checked variants of
load/store
- Add new overloads of `joint_matrix_load` and `joint_matrix_store`
where the offsets are separated from the base pointer and added as
separate arguments. I kept the same name as the expectation is to remove
the regular variants once the new ones are used instead.
- Add restrictions on both the regular and the offset
`joint_matrix_load/store` on PVC since in the current implementation, no
runtime checks are added as they are expensive. The fall back to 1d
load/store is done using a flag instead.