[SYCL][ESIMD][E2E] Fix linear test on new GPU driver (#15790)
The root cause of this test failure on the new driver is that this test
by design reads beyond the bounds of the input image and uses the out of
boundary data in calculations for the output. New driver returns 0s for
all the data read beyond the image boundaries but the old driver
replicates the last read data. I am not sure if this is the driver issue
or an issue with image/accessor and therefore the safest solution that
will work on old and new driver is to replicate the behavior of the old
driver in the test so it will work both for old and new drivers