[SYCL] Don't print the header and footer in the preprocessed output. (#15634)
This is needed to support compilation with `ccache`.
When compiling with `ccache`, the integration header and footer
shouldn't be printed in the pre-processed output to not confuse
`ccache`. In order to do that we added a new option called
`include-internal-footer` to include the integration footer. For
consistency the `include` option, that was including the integration
header was renamed `-include-internal-header`.