[Doc] Update docs to clarify cache-related environment variables. (#15795)
We have two caches: in-memory and persistent. For some cache-related
environment variables, it is not explicitly clear whether they are
applicable for in-memory cache or persistent. So, this PR clarifies the
description of cache-related environment variables.

Side note: I see that we never use `SYCL_CACHE_EVICTION_DISABLE`,
`SYCL_CACHE_MAX_SIZE`, and `SYCL_CACHE_THRESHOLD` in the code base, yet
we have them in the documentation. I think, for now, we can just leave
them as it is since I'm working on implementing cache eviction and I can
reuse them in future.