[SYCL] Deprecate `[[intel::reqd_sub_group_size]]` (#15798)
That particular spelling had been introduced as part of an Intel extension to support sub-groups in SYCL 1.2.1 before they became a core feature of SYCL 2020.

The extension has been deprecated for a while and no one should use that legacy spelling anymore. The official SYCL 2020 spelling should be used instead (with `sycl::` namespace).