[SYCL] Rewrite some E2E tests into unit-tests (#15877)
The main idea behind this (and other PRs I will eventually submit) is
that we shouldn't use `SYCL_UR_TRACE` in our unit-tests unless
absolutely necessary (like testing shutdown sequence).

The reason behind this is the fact that we have unit-tests
infrastructure which is more than capable of ensuring that we called the
right UR APIs with the right arguments. There is no need to waste our CI
time on compiling and running the same test on different devices if we
can easily test the same via our UR mock infrastructure.