Avoid potential ambiguity in UrReturnHelper
If UrReturnHelper's operator() is explicitly called as
operator()<T>(...), there is a potential for ambiguity when the
specified RetType and the inferred T are the same: this is ambiguous
with the version of operator() where only T is a template parameter,
and T is specified explicitly. We already have code that explicitly
calls operator()<T>(...), so prevent this from becoming a problem.