[Driver][SYCL][NFC] Updates to sync up with upstream
To better match up with what is occurring in the upstream, sync up the
changes in the driver.  This sync-up should make it easier to push
future changes upstream and reduce confusion as items are pulled down.

Changes include updates to function names, comment updates and some
restructuring of how options are passed to the clang -cc1 step.