[SYCL][NFC] Detach unit-tests naming from UR (#14815)
This is a follow-up from #14145. Our unit tests have mock classes
corresponding to device images and other entries our compiler generates.
They are all named using `Ur` prefix, but they have nothing to do with
Unified Runtime, they are part of our compiler interface.

This patch renames associated files and classes so they do not refer to
`Ur` anymore.
This is a mechanical find-and-replace change.