[NFCI][SYCL] Introduce `oneapi::experimental::detail::property_base`
We're planning a refactoring of the compile-timer properties design. One
part of it would be to eliminate `property_value` class template and
instead use inheritance to simplify properties' creation/implement
infrastructure for properties lists.

This PR is the first change in that direction. It doesn't bring new
functionality but introduces the necessary base class in all existing
properties and ensures (via static_assert) that any new property added
will have to make use of it as well. Having it in-tree already will
simplify subsequent PRs.