[SYCL][E2E] Avoid unused argument warning for imf_fp16_trivial_test (#16049)
The imf_fp16_trivial_test has a clangxx invocation with only object
files. On Windows, this invocation may have /EHsc, which in this case
will be ignored. This issues a warning, which in turn becomes an error
when -Werror is set. To avoid this, we ignore unused-argument warnings
for this specific invocation.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>