[Driver][SYCL][NFC] Fix memory sanitizer issue for AOT (#16063)
When using AOT for GPU, we will setup implied options for PVC targets.
The strings that are formed need to be added to the vector to be
processed later. One of these strings was a StringRef so the pointed to
location can go away. Make a copy of the string to satisfy the
sanitizer.