[SYCL] Add indeterminate constructor to work group memory interface (#16003)
Implement the changes made to the work group memory spec in this PR:
https://github.com/intel/llvm/pull/15933.
Essentially, the default constructor is removed from the spec and
instead a constructor that takes an indeterminate argument is made
available in order to emphasize that any operation on such a work group
memory object is undefined behavior except for assigning to it another
work group memory object. Because the frontend needs special types to
have a default constructor, we make it private instead and when its time
for the frontend to call it, simply override the access specifier
temporarily.