[SYCL][NFCI] Unify queue submit paths (#15776)
The submit method on queue has caused multiple divergences throughout
its lifetime, when additional information was required for new
functionality. However, that design means an exponential growth in
functions every time a new optional argument is needed. To battle this,
this patch adds a new pimpl class with the optional submission info,
which will only be initialized if needed. This boils the submit function
paths down to one with event and one without.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>