[NFCI][SYCL] Remove most uses of `ContainsProperty`
First, `properties` itself has an interface that does exactly that.
Second, this trait uses `std::tuple` which is relatively expensive to
instantiate and that happens *on top* of the normal `properties`
instantation that doesn't use `std::tuple` anymore.

Remaining uses are related to `ConflictingProperties` which is ugly
enough to be addressed in a separate PR.