[SYCL] Fix the barrier dependency for OOO profiling tags
This commit fixes an issue where the barrier before the timestamp
enqueued for the profiling tag in an out-of-order queue did not prevent
future work from being enqueued prior to the start/end of the profiling
tag.

Since the implementation now needs a marker, a barrier and a timestamp
enqueued onto the queue, this patch also optimizes the case of profiling
tags on out-of-order queues with profiling enabled, as we can achieve
the same with just a single barrier, similar to the fallback case for
devices that do not support timestamps natively.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>