[SYCL] Implement eviction for in-memory program cache (#16062)
Fixes: CMPLRLLVM-27640, https://github.com/intel/llvm/issues/2517

The PR implements LRU cache eviction policy for in-memory program
caches.
The high-level idea is to store programs in a linked-list, called
eviction list. When the program is first added to the cache, it is also
added to the eviction list. When a program is fetched from cache, we
move the program to the end of the eviction list. So, that the programs
at the beginning of the eviction list are always least recently used.
When adding a new program to cache, we check if the size of the program
cache exceeds the threshold, if so, we evict the program from cache and
corresponding kernels from Kernel and fast kernel cache.

This PR also adds a new environment variable,
`SYCL_IN_MEM_CACHE_EVICTION_THRESHOLD` that user can use to control the
size of in-memory cache. By default, cache eviction is disabled.