[SYCL][ESIMD][E2E] Fix LSC USM store test failure (#16122)
`lsc_usm_store_u32.cpp` currently fails in syclos but passes in the
internal compiler. The reason is that `rand` returns 0 and when
`sycl::bit_cast` to `float`, it ends up as a very very small floating
point number, like `1.4e-41`. In the internal compiler, this gets
optimized to zero, probably due to unsafe fp math optimizations. It is
also zero on-device. In syclos the host remains as that small number and
ends up screwing up the correctness check because we need 0.

Just explicitly return zero when the bit-casted result is below epsilon
for the type.

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>