[SYCL][CUDA][HIP] Update images enable variable
This patch adds a variable to enable image support for HIP, and updates
the one for CUDA to use the UR naming.

SYCL images support is similar for CUDA and HIP, so it makes sense to
treat them the same, and any future work on this will focus on bindless
images rather than SYCL images.