[SYCL] Honor dependencies of empty command groups
This is a recommit of https://github.com/intel/llvm/pull/16180.
Post-commit CI had a few tests failing so the original PR was reverted.
The reason of the failures is that with proper dependencies tracking we
now try to actually create images on the device and it might fail if
device isn't capable of that. This updated version of the PR also
modifies those failing tests to skip checks on unsupported devices

Co-authored-by: Sergey Semenov <sergey.semenov@intel.com>