[L0] Set exec info for all L0 kernels in UR kernel
Currently at urKernelSetExecInfo call a flag is set only
for a single L0 kernel handle associated with one of the devices.
In multi-device case there might be multiple handles and we have to set
the flag for all of them.

Currently this issue causes SYCL "free function kernel" tests to fail
in multi-device env because indirect access flag is not set for some of the modules.