[SYCL][GRAPH] Add graph E2E tests which use free function extension (#16159)
Dynamic parameters in SYCL Graphs should be used for now with
free_function_kernels extension since then the order of kernel arguments
is well defined.

The tests are a copy of few selected ones from `Graph/Update` and
modified such that they use the extension. They live in
`Graph/Update/FreeFunctionKernels`.

Since the free function kernels are not yet implemented on CUDA, we
should keep for now both copies of the tests (with and without free
functions).