[SYCL][Test] Track amount of inter-dependencies between headers
To improve compile-time, we may one day split `sycl/sycl.hpp` into
separate headers, but we need to make sure that those headers don't
include every other header so that there is an actual benefit of using
them.

This PR introduced a test to track that certain headers which we already
consider isolated are not used from anywhere else except from
`sycl/sycl.hpp`.

Note that for now we allow extension headers to include everything and
only focusing on the core SYCL headers.