[SYCL][Doc] Add execution_domain_info extension
This extension introduces the concept of an execution domain and associated
device queries to SYCL in order to address portability issues arising from
inconsistent interpretations of the term "compute unit".

The definition of "execution domain" proposed here is intended to be more
precise than the definition of compute unit -- to provide users with certain
guarantees -- while remaining sufficiently abstract to describe multiple
hardware architectures.

Signed-off-by: John Pennycook <john.pennycook@intel.com>