[SYCL][ESIMD] Fix propagation of ESIMD attribute for inlined functions (#16193)
The internal compiler uses aggressive inlining, and we have a case where
a wrapper function that calls an ESIMD function has had the ESIMD
functions body inlined into it. The parent function does not have the
ESIMD attribute, so this causes the pass that propagates the ESIMD
attribute to fail, which ends up causing `sycl-post-link` to do the
wrong thing. It works fine when the inlining does not happen because
this pass propagates the attribute. Extend the ESIMD attribute
propagation pass to also consider functions that call ESIMD intrinsics
as ESIMD functions.

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>