Add UR_ADAPTER_INFO_VERSION query
In order to support in-code management of known failures we need a query
to differentiate between the two versions of the Level Zero adapter.
This patch adds the `UR_ADAPTER_INFO_VERSION` query which returns a
`uint32_t` which is set to return `1` for all adapters except the Level
Zero adapter when being compiled in V2 mode.