[SYCL][E2E] Initialize `arch_flag` to an empty string by default (#16200)
In an environment with no devices this variable is never set, causing a
confusing error when we try to define the `clangxx` expansion. Here we
set it to an empty string by default so that instead of exiting with a
python error we report all tests as unsupported (due to having no
devices).

This same issue is why `arch_flag` was set in build-only mode, so I
removed the line that did that too.