SPIRVReader: Support LocalSizeId (#2898)
If there is no `OpExecutionMode .. LocalSize` in the input, see if
there is an `OpExecutionModeId .. LocalSizeId` and take the value for
the `reqd_work_group_size` metadata from the referenced constants
instead.

Once `LocalSizeId` has been translated to LLVM IR, it is
indistinguishable from a (non-ID) `LocalSize` execution mode.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/d20ca5dd1c92493