[SYCL][libclc] Fix cross builds
When performing cross builds, we need native versions of various tools,
we cannot assume the cross builds that are part of the current build are
executable. LLVM provides the setup_host_tool function to handle this,
either picking up versions of tools from LLVM_NATIVE_TOOL_DIR, or
implicitly building native versions as needed. Use this in more places.

This applies the changes from LLVM #97392 and #97811 and adapts them to
DPC++, and makes the same changes in other places that are only needed
for DPC++.