[libclc] Reorganize libspirv library
This moves the libspirv libraries to their own subdirectory of libclc,
reflecting how the CLC library is organized, and how the OpenCL library
may be reorganized in the future.

The libspirv bindings are organized in a similar way to CLC, with a
single root directory and then a 'generic' directory containing the
target-agnostic implementations, as well as top-level target-specific
subdirectories to override specific builtins.

The libspirv bindings still do a lot of the CLC implementations so the
separation isn't as clean as it needs to be, but this patch should help
to keep our downstream modifications better separated from upstream
changes.

Headers that were previously 'spirv' have now been renamed to 'libspirv'
to reflect this, and to better distinguish them from the 'spirv' and
'spirv64' libclc targets. That accounts for a majority of the churn.