[SYCL] Emit error when `get_backend_info` is used under `_GLIBCXX_USE_CXX11_ABI=0`
It doesn't seem that the interfaces match the spec, so it's highly
likely that these methods will be deprecated/removed anyway, so there is
little point in fixing them to work in `_GLIBCXX_USE_CXX11_ABI=0` mode.

On the other hand, using `_GLIBCXX_USE_CXX11_ABI=0` causes segmentation
fault crashes due to ABI mismatches in runtime, so emitting compile-time
error is a better thing to do here.