SPIRVReader: Support LocalSizeHintId (#2907)
If there is no `OpExecutionMode .. LocalSizeHint` in the input, see if
there is an `OpExecutionModeId .. LocalSizeHintId` and take the value
for the `work_group_size_hint` metadata from the referenced constants
instead.

Once `LocalSizeHintId` has been translated to LLVM IR, it is
indistinguishable from a (non-ID) `LocalSizeHint` execution mode.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/420f5dd41a21d5a