[SYCL][E2E] Expand fp64 emulation testing (#16259)
Updated the test to cover more HW so that there is a less change that we
will simply skip it.

Note that FP64 emulation is only enabled by default on DG2. Therefore,
we include HW which
natively supports `fp64`, but can't include any HW without `fp64`
support except for DG2 unless
we tweak the environment to force-enable emulation on those platforms.