[SYCL][NFCI] Address static analyzer concerns (#16296)
Two changes were made:

Special `padding` field in accessors is now zero-initialized. We
construct a `host_accessor` object in our stream implementation and
static analyzer complained that some of the accessor's fields were
uninitialized. It is not like we even reference `padding` field from
anywhere, but the code change seems small, so I decided to add the
initializer.

Fixed a condition to prevent access to an array of `N` elements at index
`N`, which is outside of the array because they are 0-indexed.

---------

Co-authored-by: Marcos Maronas <marcos.maronas@intel.com>