[SYCL][Graph] Fix dyn_cgf_accessor_spv.cpp Test (#16295)
The E2E SYCL-Graph test `Update/dyn_cgf_accessor_spv.cpp` checks that
after the first execution of the graph, the accessor referencing
`bufferB` has not been written to. This is done by checking for zero on
the line

```cpp
assert(check_value(i, 0, HostDataB[i], "HostDataB"));
```

However, we never explicitly initialize `bufferB` to zero in the test,
it is still uninitialized at the point of this assert. Therefore this
check against zero is based on UB. This patch fixes the test by
initializing the buffers to zero at the beginning of the test.