[benchmarks] add explicit benchmark groups
This patch adds back the previously removed bar charts,
with one critical difference - we no longer normalize all
results to a baseline 100%, but instead, benchmarks need
to be explicitly grouped for the comparison bar charts
to be generated. The groups need to be chosen such that
the units are all the same, and the values are roughly
similar.