[SYCL] Do not use handler::addAccessorReq() (#16111)
This function might be useful only when an accessor is connected with
different
CG that was used during a construction of the accessor. Currently it’s
helpful
only in single case of reduction, and in this case the accessor can be
connected
to a right CG in the constructor. Modify the use case and remove all
calls to
addAccessorReq().