[benchmarks] pin benchmarks to specific cores
Currently we use numactl to pin the benchmark scripts to
an entire numa node. However, after some testing, this
proved to be insufficient to make all the benchmarks stable.

This patch will change the workflow to now use taskset
to pin the benchmark script to physical cores of the
first numa node, with the exception for the first four,
which are more likely to be used by the kernel for
bookkeeping work or e.g., handling interrupts.