[SYCL][NVPTX] Obey -fcuda-short-ptr when compiling SYCL for NVPTX (#15642)
This flag turns pointers to CUDA's `shared`, `const`, and `local`
address spaces into 32-bit pointers. This can potentially save on
registers used for addressing calculations.

This option was being accepted by the frontend when compiling SYCL code,
but was then reporting an error that the backend datalayout doesn't
match the expected target description. This was because the option
wasn't being caught by all parts of the toolchain, leading to
inconsistencies.

This PR allows users to pass the option if they wish. They will see a
warning that the compiler is linking against a libclc/libspirv that
hasn't been compiled with this option, but this is likely harmless since
libspirv doesn't manipulate pointers.