[SYCL][libdevice][CMake][Trivial] Fix imf-fallback missing in NVPTX
Seems like the IMF fallback libraries are not collected in the merged libdevice for NVPTX and AMDGPU, because the name of the CMake property was wrong.

I don't know why our testing does not catch this. I have seen this result in unresolved symbols earlier, but I have been sitting on this patch for a while and did not try with current trunk.