[SYCL][E2E][Subdevice] Reduce run time of test subdevice_pi.cpp
On intel cpu device, the number of subdevices was the same as cpu count.
The number could be large on a server. In `shared` subtest, all
subdevices belong to the same context and device program is built for
every subdevice. Device code compilation may take long time.
This PR reduces test time from 10s to 0.4s on 160-core ICX. When device
sanitizer is enabled, test time is reduced from ~10min to 14s.