Fix test
Update is only allowed if the kernel object is the same. Calling
ext_oneapi_get_kernel is creating a new kernel object for each
invocation, since the implementation doesn't cache kernels.

The user should not assume the caching of kernels, so must use the same
kernel objects each time for update.