[SYCL] Move SYCLLowerWGLocalMemoryPass to OptimizerEarlyEPCallback
The pass transforms __sycl_allocateLocalMemory call to access of global
variable @WGLocalMem .

Move the transform to beginning of the optimizer since the access
could enable more optimization than the function call.

In addition, intel gpu compiler has a pass to transform global variable
in addrspace(3) to alloca, we must run SYCLLowerWGLocalMemoryPass before
it.