[CI] New workflow for Nightly on sycl-rel (#16297)
We're working on setting a validation for sycl-rel branches. This PR
introduces a new workflow to run Nightly on sycl-rel-* branches.

GH actions allows to run scheduled workflows only if the workflow is on
the default branch. So we have to checkout to the current rel branch for
every job.