Add e2e test for queue::fill with a range of pattern sizes (#15991)
~~Update the UR tag to include
https://github.com/oneapi-src/unified-runtime/pull/2273 fixing
`queue::fill` for the CUDA and HIP backends. It was previously producing
incorrect outputs for any pattern size other than 1, 2, or a multiple of
4 bytes. A new optimisation is also added which speeds up the fill
greatly if the pattern equals to the first word repeated throughout
(e.g. all zeros). See the UR PR for more details.~~

_The UR tag update was collected in
https://github.com/intel/llvm/pull/16040 so now this PR only adds an e2e
test as stated below._

Add a new e2e test to validate `queue::fill` outputs for any pattern
size between 1 and 32 bytes. This test fails for CUDA and HIP before the
UR change and passes with this PR. Other backends already worked
correctly.