[SYCL] Implement "swizzle" member function on swizzles (#16353)
It brings it closer to SYCL spec and makes more consistent by having
"named" swizzle member functions (like `.hi()`/`.lo()`) and `.swizzle<...>()`
behave in a similar way, i.e. we still have a bug when doing a swizzle on an
expression tree.

Note that the whole "expression tree" machinery is not
standard-conformant and will be completely removed separately (under
`-fpreview-breaking-changes` flag).

I still want to implement this partial bugfix so that I could switch to
a unified mixin-based implementation for swizzles on `vec`/`swizzle`
classes, instead of doing preprocessor tricks with `swizzles.def` file.