[SYCL][Devops] Propagate a new secret to containers build (#16384)
Extended build containers action with an extra argument which specifies
a password that will be assigned to the `sycl_ci` user created within
containers.

For now this new secret is unused, so this changes is expected to have
no impact on our CI.

This is outlined from #16361 to improve testing for that PR: for
security reasons actions are only invoked from the default branch and
never from a PR branch. Therefore, to actually test that this secret is
properly used without errors we need to update the action first.