[SYCL] test improvement, issue 16309 (#16389)
The kernel-bundle-merge-options test was recently marked XFAIL (and
tracker #16309 opened).

The root problem is that when using the MSVC CL compiler, the tests set
the `debug_option` flag to `/DEBUG` but it is `/Zi` that gets converted
to `-g` when the device driver is called. Updating that python var and
removing the `XFAIL` from the test.